# src/main/java/com/example/busan/reservation/ReservationController.java
package com.example.busan.reservation;

import com.example.busan.auth.domain.Authorized;
import com.example.busan.auth.dto.Authentication;
import com.example.busan.reservation.dto.CancelReservationRequest;
import com.example.busan.reservation.dto.CreateReservationRequest;
import com.example.busan.reservation.dto.FindReservationRequest;
import com.example.busan.reservation.dto.ReservationResponse;
import com.example.busan.reservation.dto.UpdateReservationRequest;
import com.example.busan.reservation.service.ReservationService;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/reservations")
public class ReservationController {

    private final ReservationService reservationService;

    public ReservationController(final ReservationService reservationService) {
        this.reservationService = reservationService;
    }

    @PostMapping
    public ResponseEntity<Void> create(@RequestBody final CreateReservationRequest request) {
        reservationService.create(request);

        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

    @DeleteMapping("/{reservationId}")
    public ResponseEntity<Void> delete(@PathVariable("reservationId") final Long id,
                                       @Authorized final Authentication authentication,
                                       @RequestBody final CancelReservationRequest request) {
        reservationService.deleteById(id, authentication.email(), request);

        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
    }

    @PutMapping("/{reservationId}")
    public ResponseEntity<Void> update(@PathVariable("reservationId") final Long id,
                                       @Authorized final Authentication authentication,
                                       @RequestBody final UpdateReservationRequest request) {
        reservationService.update(id, authentication.email(), request);

        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
    }

    @GetMapping
    public ResponseEntity<Page<ReservationResponse>> findAllByCurrentLoggedInMember(
            @Authorized final Authentication authentication,
            @PageableDefault final Pageable pageable) {
        final Page<ReservationResponse> response = reservationService.findAllByCurrentLoggedInMember(
                authentication.email(), pageable);

        return ResponseEntity.ok(response);
    }

    @GetMapping("/all")
    public ResponseEntity<Page<ReservationResponse>> findAll(@PageableDefault final Pageable pageable,
                                                             final FindReservationRequest request) {
        final Page<ReservationResponse> reservations = reservationService.findAll(request, pageable);

        return ResponseEntity.ok(reservations);
    }
}