# src/main/resources/static/api/openapi3.yaml
openapi: 3.0.1
info:
  title: 회의실 대관 서비스 API
  description: 회의실 대관 서비스입니다.
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /reservations:
    get:
      tags:
      - reservations
      operationId: 자신의 회의실 예약 목록 최신 순으로 보기
      parameters:
      - name: page
        in: query
        description: 페이지는 1부터 시작 (디폴트값 1)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 페이지별 사이즈 (디폴트값 10)
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reservations-all-2055281597'
              examples:
                자신의 회의실 예약 목록 최신 순으로 보기:
                  value: "{\"content\":[{\"id\":1,\"status\":\"RESERVED\",\"cancelReason\"\
                    :null,\"startTime\":\"2024-01-13T18:38:18.68911\",\"endTime\"\
                    :\"2024-01-13T20:38:18.689115\",\"name\":\"황재현\",\"phone\":\"\
                    01012341234\",\"reservedAt\":\"2024-01-13T18:38:18.689124\",\"\
                    roomId\":1,\"roomName\":\"대회의실\",\"company\":\"요기요\"},{\"id\"\
                    :2,\"status\":\"CANCELED\",\"cancelReason\":\"쓰기 싫어졌어요..\",\"\
                    startTime\":\"2024-01-13T18:38:18.689133\",\"endTime\":\"2024-01-13T20:38:18.689134\"\
                    ,\"name\":\"황재현\",\"phone\":\"01012341234\",\"reservedAt\":\"\
                    2024-01-13T18:38:18.689137\",\"roomId\":1,\"roomName\":\"대회의실\"\
                    ,\"company\":\"토스뱅크\"}],\"pageable\":\"INSTANCE\",\"last\":true,\"\
                    totalPages\":1,\"totalElements\":2,\"first\":true,\"size\":2,\"\
                    number\":0,\"sort\":{\"empty\":true,\"sorted\":false,\"unsorted\"\
                    :true},\"numberOfElements\":2,\"empty\":false}"
    post:
      tags:
      - reservations
      operationId: 회의실 예약하기
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reservations-740649445'
            examples:
              회의실 예약하기:
                value: "{\"roomId\":1,\"startTime\":\"2023-11-10T14:00:00\",\"endTime\"\
                  :\"2023-11-10T15:00:00\"}"
      responses:
        "201":
          description: "201"
  /reservations/all:
    get:
      tags:
      - reservations
      operationId: 전체 회의실 예약 목록 최신 순으로 보기
      parameters:
      - name: page
        in: query
        description: 페이지는 1부터 시작 (디폴트값 1)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 페이지별 사이즈 (디폴트값 10)
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reservations-all1582432150'
            examples:
              전체 회의실 예약 목록 최신 순으로 보기:
                value: "{\"start\":\"2024-01-13T18:38:18.800506\",\"end\":\"2024-01-13T18:38:18.800508\"\
                  }"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reservations-all-2055281597'
              examples:
                전체 회의실 예약 목록 최신 순으로 보기:
                  value: "{\"content\":[{\"id\":1,\"status\":\"RESERVED\",\"cancelReason\"\
                    :null,\"startTime\":\"2024-01-13T18:38:18.800459\",\"endTime\"\
                    :\"2024-01-13T20:38:18.800463\",\"name\":\"황재현\",\"phone\":\"\
                    01012341234\",\"reservedAt\":\"2024-01-13T18:38:18.80047\",\"\
                    roomId\":1,\"roomName\":\"대회의실\",\"company\":\"요기요\"},{\"id\"\
                    :2,\"status\":\"CANCELED\",\"cancelReason\":\"쓰기 싫어졌어요..\",\"\
                    startTime\":\"2024-01-13T18:38:18.800473\",\"endTime\":\"2024-01-13T20:38:18.800474\"\
                    ,\"name\":\"황재현\",\"phone\":\"01012341234\",\"reservedAt\":\"\
                    2024-01-13T18:38:18.800477\",\"roomId\":1,\"roomName\":\"대회의실\"\
                    ,\"company\":\"토스뱅크\"}],\"pageable\":\"INSTANCE\",\"last\":true,\"\
                    totalPages\":1,\"totalElements\":2,\"first\":true,\"size\":2,\"\
                    number\":0,\"sort\":{\"empty\":true,\"sorted\":false,\"unsorted\"\
                    :true},\"numberOfElements\":2,\"empty\":false}"
  /reservations/{reservationId}:
    put:
      tags:
      - reservations
      operationId: 자신의 회의실 예약 수정하기
      parameters:
      - name: reservationId
        in: path
        description: 수정할 예약 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reservations-740649445'
            examples:
              자신의 회의실 예약 수정하기:
                value: "{\"roomId\":1,\"startTime\":\"2023-11-10T14:00:00\",\"endTime\"\
                  :\"2023-11-10T15:00:00\"}"
      responses:
        "204":
          description: "204"
    delete:
      tags:
      - reservations
      operationId: 자신의 회의실 예약 취소하기
      parameters:
      - name: reservationId
        in: path
        description: 취소할 예약 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reservations-reservationId-361069914'
            examples:
              자신의 회의실 예약 취소하기:
                value: "{\"reason\":\"쓰기 싫어요\"}"
      responses:
        "204":
          description: "204"
...
components:
  schemas:
...
    reservations-740649445:
      type: object
      properties:
        startTime:
          type: string
          description: 시작 시각
        endTime:
          type: string
          description: 종료 시각
        roomId:
          type: number
          description: 예약할 회의실 ID
...
    reservations-all1582432150:
      type: object
      properties:
        start:
          type: string
          description: 시작일(포함)
        end:
          type: string
          description: 마지막일(포함)
...
    reservations-reservationId-361069914:
      type: object
      properties:
        reason:
          type: string
          description: 취소 사유
    reservations-all-2055281597:
      type: object
      properties:
        number:
          type: number
          description: 현재 페이지
        size:
          type: number
          description: 한 페이지당 사이즈
        numberOfElements:
          type: number
          description: 현재 응답 데이터 요소 개수
        last:
          type: boolean
          description: 마지막 페이지인지
        totalPages:
          type: number
          description: 전체 페이지 개수
        first:
          type: boolean
          description: 첫 페이지인지
        content:
          type: array
          items:
            type: object
            properties:
              phone:
                type: string
                description: 예약자 휴대폰 번호
              name:
                type: string
                description: 예약자 성함
              company:
                type: string
                description: 예약자 회사
              startTime:
                type: string
                description: 시작 시각
              endTime:
                type: string
                description: 종료 시각
              id:
                type: number
                description: 예약 ID
              reservedAt:
                type: string
                description: 예약 일시
              cancelReason:
                type: string
                description: 취소 상태일 경우 취소 이유(취소가 아니면 null)
              roomName:
                type: string
                description: 회의실 이름
              roomId:
                type: number
                description: 회의실 ID
              status:
                type: string
                description: 예약 상태
        totalElements:
          type: number
          description: 전체 DB 데이터 요소 개수
        empty:
          type: boolean
          description: 빈 데이터인지
...